% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-elb.R
\name{elbSummary}
\alias{elbSummary}
\title{Generate a daily, weekly, whatever summary of web traffic}
\usage{
elbSummary(days, before.date = Sys.Date(), send = TRUE)
}
\arguments{
\item{days}{Integer number of days to aggregate}

\item{before.date}{days before this day. Default is today (i.e. with days=1,
give the report for yesterday)}

\item{send}{Logical: send messages to Slack?}
}
\description{
Generate a daily, weekly, whatever summary of web traffic
}
