% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow-requests.R
\name{slowRequestReport}
\alias{slowRequestReport}
\title{Send a useful summary of slow request timings from superadmin to slack}
\usage{
slowRequestReport(minutes = 60, threshold = 119, max.reports = 5,
  send = TRUE, channel = "#systems", ...)
}
\arguments{
\item{minutes}{How big of a window into the past do we care?}

\item{threshold}{How many seconds is "too slow"? Passed to superadmin query}

\item{max.reports}{How many reports should we include in the Slack message?}

\item{send}{Should we send to Slack at all? If not, just prints to screen}

\item{channel}{Which Slack channel should the report be sent to?}

\item{...}{Additional arguments passed to \code{slack()}}
}
\value{
If \code{send}, the Slack API response object. If not, then whatever \code{print}' returns.
}
\description{
Send a useful summary of slow request timings from superadmin to slack
}
