% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{with_slack_errors}
\alias{with_slack_errors}
\title{Context to capture errors and report them in slack}
\usage{
with_slack_errors(expr)
}
\arguments{
\item{expr}{Code to evaluate in the context}
}
\value{
The output of expr. If any error is hit, the error message will go to
Slack.
}
\description{
Context to capture errors and report them in slack
}
