% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-reported.R
\name{supportReport}
\alias{supportReport}
\title{Support Reports}
\usage{
supportReport(date = "yesterday", subtitle = paste("Date range:", date),
  ...)
}
\arguments{
\item{date}{string Pivotal date query}

\item{subtitle}{string to use as "pretext" in Slack attachment}

\item{...}{additional arguments passed to \code{slack}}
}
\value{
A \code{httr} response from the slack POST.
}
\description{
Support Reports
}
